+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### Changelog
 
+ - 1/25/2014 v0.1.2: Hide visitor prices.
+
  - 1/16/2014 v0.1.1: Update date parsing string.  The Middlebury Menu changed how it displays it's date.